refactor (handleController.ts): rename upload --> handle

+++ src/controllers/handleController.ts
@@ -5,30 +5,18 @@ import fs from 'fs';
 import { cleanOutDirectory } from '../services/fileService';
 import { uploadDir, outDir } from '../config/multerConfig';
 
-export const uploadSolidity = (req: Request, res: Response, next: NextFunction) => {
-    if (req.file) {
-        const filePath = path.join(uploadDir, req.file.originalname);
+export const handleSolidity = async (req: Request, res: Response, next: NextFunction) => {
+    if (req.file && req.file.originalname.endsWith('.sol')) {
+        const filePath = path.join(SolDir, req.file.originalname);
         const fileNameWithoutExtension = req.file.originalname.replace(/\.[^/.]+$/, "");
 
         const { privateKey, rpcUrl } = req.body;
 
-        const forgeCommand = `forge create ${filePath}:${fileNameWithoutExtension} --rpc-url ${rpcUrl} --private-key ${privateKey}`;
 
-        exec(forgeCommand, (error, stdout, stderr) => {
-            if (error) {
-                console.error(`Error executing command: ${error.message}`);
-                res.status(500).send(`Error executing command: ${error.message}`);
-                return;
-            }
+        try {
+            await deploySolidity(filePath, fileNameWithoutExtension, rpcUrl, privateKey, res);
 
-            if (stderr) {
-                console.error(`Error output: ${stderr}`);
-                res.status(500).send(`Command failed: ${stderr}`);
-                return;
-            }
-
-            res.send(`Command output: ${stdout}`);
-
+            // Delete the file after successful deployment
             fs.unlink(filePath, (err) => {
                 if (err) {
                     console.error(`Error deleting file: ${err.message}`);
@@ -37,8 +25,18 @@ export const uploadSolidity = (req: Request, res: Response, next: NextFunction) 
                 }
             });
 
+            // Clean out the directory
             cleanOutDirectory(outDir);
-        });
+        } catch (error: unknown) {
+            if (error instanceof Error) {
+                console.error(`Deployment failed: ${error.message}`);
+                res.status(500).send(`Deployment failed: ${error.message}`);
+            } else {
+                console.error('An unknown error occurred during deployment');
+                res.status(500).send('An unknown error occurred during deployment');
+            }
+        }
+
     } else {
         res.status(400).send('No file uploaded or file is not .sol');
     }
